--- a/build.sbt
+++ b/build.sbt
@@ -89,6 +89,7 @@ lazy val benchmark = Project("benchmark", file("benchmark"))
       "io.circe" %% "circe-core" % "0.11.1",
       "io.circe" %% "circe-generic" % "0.11.1",
       "com.typesafe.play" %% "play-json" % "2.7.4",
+      "com.fasterxml.jackson.module" %% "jackson-module-scala" % "2.9.9",
     )
   )