String parsing dominates parsing times so far for real world input (that isn't particularly number heavy). So, here's a first attempt to improve performance.

The speed-ups mostly come from
 * no explicit bounds checking
 * removing tricks
 * some code reorganization

Before:

```
[info] GithubIssuesJsonAST.readCirce                github-akka-issues.json  thrpt    5  2465.716 ± 191.574  ops/s
[info] GithubIssuesJsonAST.readCirce                  big-string-array.json  thrpt    5  3375.916 ± 317.283  ops/s
[info] GithubIssuesJsonAST.readPlayJson             github-akka-issues.json  thrpt    5  2002.679 ± 159.931  ops/s
[info] GithubIssuesJsonAST.readPlayJson               big-string-array.json  thrpt    5  2790.919 ±  59.378  ops/s
[info] GithubIssuesJsonAST.readSprayJsonFromBytes   github-akka-issues.json  thrpt    5  1048.052 ±  12.170  ops/s
[info] GithubIssuesJsonAST.readSprayJsonFromBytes     big-string-array.json  thrpt    5  1353.650 ±  27.423  ops/s
[info] GithubIssuesJsonAST.readSprayJsonFromString  github-akka-issues.json  thrpt    5  1202.670 ±  28.346  ops/s
[info] GithubIssuesJsonAST.readSprayJsonFromString    big-string-array.json  thrpt    5  1507.546 ±  19.228  ops/s

[info] GithubIssuesCaseClassReading.readSprayJsonViaASTPartial  thrpt    5  1141.810 ± 3.432  ops/s
```

After:

```
[info] GithubIssuesJsonAST.readSprayJsonFromBytes   github-akka-issues.json  thrpt    5  2321.173 ±  74.728  ops/s
[info] GithubIssuesJsonAST.readSprayJsonFromBytes     big-string-array.json  thrpt    5  4250.720 ± 187.668  ops/s
[info] GithubIssuesJsonAST.readSprayJsonFromString  github-akka-issues.json  thrpt    5  2181.839 ±  60.334  ops/s
[info] GithubIssuesJsonAST.readSprayJsonFromString    big-string-array.json  thrpt    5  3447.510 ±  27.173  ops/s

[info] GithubIssuesCaseClassReading.readSprayJsonViaASTPartial  thrpt    5  2142.250 ± 99.059  ops/s
```

I.e. about 2x-3x faster for byte input. So, string parsing has been improved by ~ 3x which leads to a 2x improvement for the test case (88k json from github API).

As `ParserInput` needs to be changed, the change won't be simply binary compatible. I'll probably deprecate the current parser infrastructure and duplicate it by simpler entry-points that don't leak so many implementation details for common cases.