@sirthias 2019-07-30
Awesome stuff, @jrudolph!
These are great improvements!!
The old "tricks" really don't pull their weight, especially on the newer JVMs.