@sirthias 2019-07-30
One thing that might be worthwhile exploring to speed up string parsing even more is unrolling the `parseOneFast` loop to handle 4 characters in a single loop iteration.
I could imagine that we getter instruction level parallelism this way.

The beginning of this blog post here: https://chadaustin.me/2017/05/writing-a-really-really-fast-json-parser/ discusses this effect for JSON parsing.

In the end it's all about reducing the average parsing cycles per character, so any parallelization that we can get will help.

Another thing could be sth similar to this line of _borer's_ JSON parser:
https://github.com/sirthias/borer/blob/master/core/src/main/scala/io/bullet/borer/json/JsonParser.scala#L417

Decoupling the character copying from the rest of the loop had an incredible effect on performance in borer's case. I suspect because the CPU can then freely move those instructions between everything else and make use of otherwise unused silicon capacity, so a lot of it gets essentially moved out of the critical path and costs "nothing" any more...