@plokhotnyuk 2019-07-30
@sirthias I have found that C2 can do unrolling for some simple loops, and for [my case](https://github.com/plokhotnyuk/jsoniter-scala/commit/2a8cad7e550833d714b539a64413cc84e5302323#diff-7ee06016142354570a183b31a6f45075) only GraalVM required for manual unrolling