@sirthias 2019-07-30
Yeah, the JVM _might_ be able to do it itself.
But this is hard to control and small changes might break the JVMs ability to do it.
So manual unrolling for this very hot code path appears to be worth the effort.