@plokhotnyuk 2019-07-30
Here are throughput scores of benchmarks which parse a string of 128 ASCII chars on the same Intel Core i7 CPU but on different versions of JDK and GraalVM:
![image](https://user-images.githubusercontent.com/890289/62106793-6b898280-b2a6-11e9-872f-81ceb9f3d8ae.png)

It is much harder to keep them high on different hardware.