@jrudolph 2019-07-30
> You can probably improve a lot in these benchmarks when you bring-your-own-buffer.

Caching the buffer in a thread local between parsing runs is another 10-20% win indeed.