@sirthias 2019-07-30
> Caching the buffer in a thread local between parsing runs is another 10-20% win indeed.

Interesting. In my tests with borer adding a ThreadLocal cache of the char buffer somehow resulted in a slight performance _decrease_, at least on graal.
This perf work really is a complete trial and error process... unfortunately.