@jrudolph 2019-07-30
>  This perf work really is a complete trial and error process... unfortunately.

It is, I'm also still looking for a more predictable workflow. I didn't test extensively on Graal, from what I've seen, GraalCE was slower than hotspot and JDK11 was a bit faster than JDK8.